Change function signature from float to double to match type get_doub…

+++ core/io/stream_peer.cpp
@@ -310,7 +310,7 @@ float StreamPeer::get_float() {
 	return decode_float(buf);
 }
 
-float StreamPeer::get_double() {
+double StreamPeer::get_double() {
 
 	uint8_t buf[8];
 	get_data(buf, 8);
